[feat/#99] 루트 레이아웃 책임 분리 및 hooks 분리
