+++ src/app/_layout.tsx
@@ -3,17 +3,14 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { Stack, useSegments } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 import "react-native-reanimated";
-import * as SecureStore from "expo-secure-store";
-import { useEffect } from "react";
 import { StyleSheet, useWindowDimensions, View } from "react-native";
 
 import "../../global.css";
 
 import { BottomSheetProvider } from "@/app/_providers";
 import { ChatFloatingButton, ChatSheet, useChatStore } from "@/features/chat";
-import { useAuthStore } from "@/features/kakao-login";
+import { useRestoreAuthToken } from "@/features/kakao-login";
 import { useShelfDetailStore } from "@/features/view-shelf-detail";
-import { AUTH_TOKEN_KEY } from "@/shared/config/auth-storage";
 import { semanticColors } from "@/shared/config/tokens";
 import { useColorScheme } from "@/shared/lib/hooks/use-color-scheme";
 
@@ -24,64 +21,67 @@ export const unstable_settings = {
 };
 
 export default function RootLayout() {
-  const { height } = useWindowDimensions();
-  // 시트 snapPoint가 75%이므로 나머지 25% + 여유분을 덮는다
-  const overlayHeight = height * 0.3;
+  return (
+    <QueryClientProvider client={queryClient}>
+      <AppProviders>
+        <RootLayoutContent />
+      </AppProviders>
+    </QueryClientProvider>
+  );
+}
+
+/**
+ * Wraps the app with required providers (BottomSheet, Theme).
+ * Keeps provider setup isolated from layout logic.
+ */
+function AppProviders({ children }: { children: React.ReactNode }) {
   const colorScheme = useColorScheme();
+
+  return (
+    <BottomSheetProvider>
+      <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
+        {children}
+      </ThemeProvider>
+    </BottomSheetProvider>
+  );
+}
+
+/**
+ * Root layout content: handles navigation, floating UI, and overlays.
+ */
+function RootLayoutContent() {
+  const { height } = useWindowDimensions();
+  const overlayHeight = height * 0.3;
+
+  // Restore auth token on app startup
+  useRestoreAuthToken();
+
+  // UI state
+  const segments = useSegments();
+  const isInAuth = segments[0] === "(auth)";
   const openChat = useChatStore((s) => s.open);
   const isChatOpen = useChatStore((s) => s.isOpen);
   const isShelfDetailOpen = useShelfDetailStore((s) => s.selectedType !== null);
-  const segments = useSegments();
-  const isInAuth = segments[0] === "(auth)";
-
-  const setAuth = useAuthStore((s) => s.setAuth);
-  const clearAuth = useAuthStore((s) => s.clearAuth);
-
-  // Restore token on app startup
-  useEffect(() => {
-    const restoreToken = async () => {
-      try {
-        const token = await SecureStore.getItemAsync(AUTH_TOKEN_KEY);
-        if (token) {
-          // TODO: Call backend /auth/me to get user info and validate token
-          // For now, we'll just set a placeholder user - backend should return actual user data
-          setAuth({ id: "", nickname: "", profileImage: "" }, token);
-        } else {
-          clearAuth();
-        }
-      } catch (error) {
-        console.error("Failed to restore token:", error);
-        clearAuth();
-      }
-    };
-
-    restoreToken();
-  }, [setAuth, clearAuth]);
-
-  const _isOnScan = (segments as string[]).includes("scan");
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <BottomSheetProvider>
-        <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
-          <Stack screenOptions={{ headerShown: false }}>
-            <Stack.Screen name="(auth)" />
-            <Stack.Screen name="(protected)" />
-          </Stack>
+    <>
+      <Stack screenOptions={{ headerShown: false }}>
+        <Stack.Screen name="(auth)" />
+        <Stack.Screen name="(protected)" />
+      </Stack>
 
-          {!isChatOpen && !isInAuth && !isShelfDetailOpen && (
-            <ChatFloatingButton onPress={openChat} />
-          )}
+      {/* Chat floating button: hidden when chat is open, in auth flow, or shelf detail is open */}
+      {!isChatOpen && !isInAuth && !isShelfDetailOpen && <ChatFloatingButton onPress={openChat} />}
 
-          <ChatSheet />
-          {/* 채팅창이 열려있을 때 시트 아래 노출 영역을 아이보리로 덮는 오버레이 */}
-          {isChatOpen && (
-            <View pointerEvents="none" style={[styles.chatBgOverlay, { height: overlayHeight }]} />
-          )}
-          <StatusBar style="auto" />
-        </ThemeProvider>
-      </BottomSheetProvider>
-    </QueryClientProvider>
+      <ChatSheet />
+
+      {/* Overlay to cover content below when chat sheet is open */}
+      {isChatOpen && (
+        <View pointerEvents="none" style={[styles.chatBgOverlay, { height: overlayHeight }]} />
+      )}
+
+      <StatusBar style="auto" />
+    </>
   );
 }
