+++ src/features/kakao-login/index.ts
@@ -2,6 +2,7 @@ export type { KakaoLoginResponse } from "./api/kakao-auth";
 export { kakaoLogin, kakaoLogout } from "./api/kakao-auth";
 export { useAuthStore } from "./model/store";
 export { useKakaoLogin } from "./model/use-kakao-login";
+export { useRestoreAuthToken } from "./model/use-restore-auth-token";
 export { KakaoLoginButton } from "./ui/KakaoLoginButton";
 export { KakaoSymbol } from "./ui/KakaoSymbol";
 export { OrDivider } from "./ui/OrDivider";
